Use double-precision SetPosX in DrawUniteList unite-panel labels
The debug target computes the unite-panel label SetPosX in double precision
(fsub/fmadd on the 0.5 and panel->x), narrowing only at the call; our early
float casts forced single ops. Dropping them matches the target, lifting
DrawUniteList 67.04->67.48 and the unit 71.68->71.75.

+++ src/menu_cmd.cpp
@@ -2414,9 +2414,9 @@ void CMenuPcs::DrawUniteList()
 		const int itemId = caravan->m_commandListExtra[s_UniteTop[i]];
 		const char* text = GetStrikeListName(itemId);
 		const float width = static_cast<float>(font->GetWidth(text));
-		font->SetPosX((static_cast<float>(panel->width) - width) *
-		                  static_cast<float>(kCmdMenuHalfD) +
-		              static_cast<float>(panel->x));
+		font->SetPosX(static_cast<float>((static_cast<double>(panel->width) - width) *
+		                  kCmdMenuHalfD +
+		              static_cast<double>(panel->x)));
 		font->SetPosY(static_cast<float>(((static_cast<float>(panel->height) - kCmdMenuTextLineHeightD) *
 		                                      kCmdMenuHalfD +
 		                                  static_cast<float>(panel->y)) -
